+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.2] - 2026-Apr-14
+- systemd daemon timer for rollback logic
+- reading pending.json file for restore as needed
+- fleshing out `apply` logic, partway complete
+    - dry run and validation
+    - ruleset snapshotting
+    - systemd timer creation
+
 ## [0.1.1] - 2026-Apr-14
 - commit.go package creation
 - rollback/snapshot capture bones